requirements.txt
pdfminer.six
streamlit==1.35.0
python-dotenv==1.0.1
python-docx
markdown
htmldocx
chromadb==0.5.0
llama-index==0.10.44
llama-index-core==0.10.44
llama-index-vector-stores-chroma==0.1.9